+++ react/utils/routes.ts
@@ -8,6 +8,7 @@ import { generateExtensions } from './blocks'
 import { fetchWithRetry } from './fetch'
 import { parseMessages } from './messages'
 import { isEnabled } from './flags'
+import { DeviceInfo } from './withDevice'
 
 const parsePageQueryResponse = (
   page: PageQueryResponse
@@ -133,11 +134,13 @@ export const fetchServerPage = async ({
   path,
   query: rawQuery,
   workspace,
+  deviceInfo,
 }: {
   path: string
   query?: Record<string, string>
   fetcher: GlobalFetch['fetch']
   workspace?: string
+  deviceInfo?: DeviceInfo
 }): Promise<ParsedServerPageResponse> => {
   const url = getRelativeURLWithQuery({
     path,
@@ -145,6 +148,9 @@ export const fetchServerPage = async ({
       ...rawQuery,
       ...(workspace ? { workspace } : {}),
       __pickRuntime: runtimeFields,
+      ...(deviceInfo && {
+        __device: deviceInfo.type,
+      }),
     },
   })
   const page: ServerPageResponse = await fetchWithRetry(