+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Blocks for different devices are loaded lazily if the screen resizes to their breakpoints.
+- Provides deviceInfo on runtime context.
 
 ## [8.113.0] - 2020-09-06
 ### Added